Fix cm versioning issue, run gofmt
When creating an cm issuer we were using the wrong versioning, i.e
`v1alpha2`. Most of the certmanager APIs have been V1 for a while,
therefore I think it's safe to say we can default to that. If not we
can always just get the CRD for the issuer and scrape the versioning
from that. (possibly best for a follow up issue).

Signed-off-by: Andrew Stoycos <astoycos@redhat.com>